Complete Spring Tool Suite Project for Film rental Application
Project is to build a Spring Application that uses JPA, Spring MVC, REST Web Service with an AngularJS front end.Project is about to complete rental store application. That application will be used by the rental shop in order. The rental shop is determined by the login. This is no GUI, so to test it we have used REST clients using cURLou Chrome Dev environment